requirements.txt
pandas
numpy
requests
beautifulsoup4
openai
python-dotenv
